requirements_app.txt
SQLAlchemy==1.3.17
Flask==2.1.1
python-dateutil==2.8.1
uWSGI==2.0.18
psycopg2==2.9.1
redis==3.5.3
